+++ src/main/java/inu/spp/cryptocoffee/user/UserRepository.java
@@ -1,10 +1,8 @@
-package inu.spp.cryptocoffee.domain.repository;
+package inu.spp.cryptocoffee.user;
 
-import inu.spp.cryptocoffee.domain.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
 }
